jakies bledy, jutro ogarne
